Change js structure
