+++ public/js/eventos.js
@@ -1,93 +1,10 @@
+import { testDishesStriked, finishDish, setFinishDishValue } from "./modules/dishesFunctions.js";
+import { setDateMinAttributeOnForm } from "./modules/datePickerFunctions.js";
+import { updateAddList, resetOrder, saveNewOrder, changeActionShowOrderForm } from "./modules/orderFunctions.js";
+
 "use strict"; 
 
-/**
- * This function updates the text of an HTML element based on the value of a sibling element.
- */
-function finishDish() {       
-    if(this.className === "finished") {
-        document.getElementById(this.id).setAttribute("class", "finished lineThrough");
-    }
-    else {
-        document.getElementById(this.id).setAttribute("class", "finished");
-    }           
-}
-
-
-/**
- * The function toggles the value of a checkbox input's previous sibling input between 0 and 1.
- */
-function setFinishDishValue() {
-    let finishedValue = document.getElementById(this.id).previousElementSibling;
-    finishedValue.value = finishedValue.value == 1 ?  0 : 1;
-}
-
-
-/**
- * The function marks completed items orders as finished by striking through their original text.
- */
-function testDishesStriked() {    
-    let itemsFinished = document.getElementsByClassName("item_finished");
-    let finished = document.getElementsByClassName("finished");
-
-    for (let index = 0; index < itemsFinished.length; index++) {        
-        if(itemsFinished[index].value == 1) finished[index].setAttribute("class", "finished lineThrough");                 
-    }
-}
-
-
-/**
- * Functions in new order view
- * 
- *  - resetOrder()
- *  - saveNewOrder()
- */
-
-// Reset New Order form
-function resetOrder() {
-    document.getElementById("new_order_form").action = "/orders/order/resetOrder";
-}
-
-// Save a new order
-function saveNewOrder() {
-    document.getElementById("new_order_form").action = "/orders/order/save";
-}
-
-
-/** Disable date before current day in forms with 'input type date' fields */
-function setDateMinAttributeOnForm() {
-    const today = new Date();
-    let day = today.getDate();
-    let mounth = today.getMonth() + 1;        
-    
-    // Add "0" before current day and month if they have only one digit (ex. "2" will be "02")
-    if(day.toString().length == 1) day = "0" + day;
-    if(mounth.toString().length == 1) mounth = "0" + mounth;
-        
-    const minDate = today.getFullYear() + "-" + mounth + "-" + day;   
-        
-    // Set the min attribute of the input field to today's date
-    document.querySelector('input[type="date"]').min = minDate;    
-}
-
-/** Change action attribute in order view when click 'update' button */
-function changeActionShowOrderForm() {         
-    let formElement = document.getElementById('show_order_form')
-
-    formElement.action = "/admin/comandas/update";
-    formElement.submit();
-}
-
-/** Change the acction attrib. in 'add_to_order_form' form */
-function updateAddList() {
-    document.getElementById('new_order_form').action = "/admin/comandas/updateAddList";
-}
-
-
-window.onload = function() {
-    /* setTimeout(function() {
-        window.location.reload(true);
-    }, 10000); */
-    
+window.onload = function() {        
     /** Test for striked dishes */
     testDishesStriked();
         
@@ -152,25 +69,5 @@ window.onload = function() {
 				}
 			});
 		});
-	}
-
-    /* let testCookies = function() {
-        //document.cookie = "test=1; expires=Thu, 23 Jun 2024 17:50:01 Europe/Madrid;";
-
-        let cookies = new Map();
-        let all = document.cookie;
-        let list = all.split('; ');
-
-        for(let cookie of list) {
-            let p = cookie.indexOf('=');
-            let name = cookie.substring(0, p);
-            let value = cookie.substring(p + 1);
-            value = decodeURIComponent(value);
-            cookies.set(name, value);
-        }
-
-        console.log(cookies);
-    }
-
-    testCookies(); */
+	}    
 }
